lto = true

+++ src/mem/mem_rchr_double.rs
@@ -397,11 +397,7 @@ fn _rchr_c16(buf_ptr: *const u8, c16: u128, start_ptr: *const u8) -> Option<usiz
         & !v
         & 0x8080_8080_8080_8080_8080_8080_8080_8080_u128;
     if bits != 0 {
-        Some(
-            plus_offset_from(buf_ptr, start_ptr) + 16
-                - 1
-                - ((bits as u128).leading_zeros() / 8) as usize,
-        )
+        Some(plus_offset_from(buf_ptr, start_ptr) + 16 - 1 - (bits.leading_zeros() / 8) as usize)
     } else {
         None
     }
@@ -421,11 +417,7 @@ fn _rchr_c8(buf_ptr: *const u8, c8: u64, start_ptr: *const u8) -> Option<usize> 
     let v = v0 ^ c8;
     let bits = v.wrapping_sub(0x0101_0101_0101_0101_u64) & !v & 0x8080_8080_8080_8080_u64;
     if bits != 0 {
-        Some(
-            plus_offset_from(buf_ptr, start_ptr) + 8
-                - 1
-                - ((bits as u64).leading_zeros() / 8) as usize,
-        )
+        Some(plus_offset_from(buf_ptr, start_ptr) + 8 - 1 - (bits.leading_zeros() / 8) as usize)
     } else {
         None
     }
@@ -445,11 +437,7 @@ fn _rchr_c4(buf_ptr: *const u8, c4: u32, start_ptr: *const u8) -> Option<usize> 
     let v = v0 ^ c4;
     let bits = v.wrapping_sub(0x0101_0101_u32) & !v & 0x8080_8080_u32;
     if bits != 0 {
-        Some(
-            plus_offset_from(buf_ptr, start_ptr) + 4
-                - 1
-                - ((bits as u32).leading_zeros() / 8) as usize,
-        )
+        Some(plus_offset_from(buf_ptr, start_ptr) + 4 - 1 - (bits.leading_zeros() / 8) as usize)
     } else {
         None
     }
@@ -469,11 +457,7 @@ fn _rchr_c2(buf_ptr: *const u8, c2: u16, start_ptr: *const u8) -> Option<usize> 
     let v = v0 ^ c2;
     let bits = v.wrapping_sub(0x0101_u16) & !v & 0x8080_u16;
     if bits != 0 {
-        Some(
-            plus_offset_from(buf_ptr, start_ptr) + 2
-                - 1
-                - ((bits as u16).leading_zeros() / 8) as usize,
-        )
+        Some(plus_offset_from(buf_ptr, start_ptr) + 2 - 1 - (bits.leading_zeros() / 8) as usize)
     } else {
         None
     }
